+++ api/src/main/java/org/pmiops/workbench/dataset/DataSetDao.java
@@ -1,4 +1,4 @@
-package org.pmiops.workbench.db.dao;
+package org.pmiops.workbench.dataset;
 
 import java.util.List;
 import org.pmiops.workbench.db.model.DataSet;
